+++ flashcard-service/src/main/java/com/revature/FlashcardServiceApplication.java
@@ -1,0 +1,13 @@
+package com.revature;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FlashcardServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FlashcardServiceApplication.class, args);
+	}
+
+}
