+++ flashcard-service/src/test/java/com/revature/FlashcardServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.revature;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FlashcardServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
