+++ quiz-service/src/main/java/com/revature/models/Difficulty.java
@@ -1,0 +1,6 @@
+package com.revature.models;
+
+public enum Difficulty {
+
+	TRIVIAL, EASY, MEDIUM, HARD, IMPOSSIBLE, RIDICULOUS
+}
