+++ quiz-service/src/main/java/com/revature/models/Flashcard.java
@@ -1,0 +1,34 @@
+package com.revature.models;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Entity
+@Getter @Setter @NoArgsConstructor @AllArgsConstructor @EqualsAndHashCode @ToString
+public class Flashcard {
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private int id;
+
+	private String question;
+	private String answer;
+	private String name;
+	private Difficulty difficulty;
+	
+	@ManyToOne(fetch = FetchType.EAGER)
+	@JoinColumn(name = "topic_id")
+	private Topic topic;
+}
