+++ quiz-service/src/main/resources/application.properties
@@ -1,0 +1,13 @@
+server.port=8080
+spring.application.name=quiz-service
+
+# Database Credentials
+spring.datasource.url=jdbc:h2:mem:memdb
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.h2.console.enabled=true
+
+# JPA Settings
+spring.jpa.show-sql=false
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
