+++ quiz-service/src/test/java/com/revature/QuizServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.revature;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class QuizServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
