--- src/main/java/com/revature/QuizAppApplication.java
@@ -1,13 +1,0 @@
-package com.revature;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class QuizAppApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(QuizAppApplication.class, args);
-	}
-
-}
